Implement {Asset,}Class<T> for unit type
Rust has some issues with cohering closures into `Class<Foo>`, seemingly
because of the arguments' lifetime.

This prevents passing a `|_| {}` when the class is irrelevant.
It is possible to pass `|_: &mut Foo| {}`, but it becomes very noisy.
Using the unit type `()` has the advantage of not introducing any
lifetimes, so it should be possible to use it when we don't want to
specify a class.